Add cards to search
