+++ src/components/card/index.ts
@@ -1,3 +1,5 @@
 import Card from './card';
+import { getCardProps } from './utils';
 
 export default Card;
+export { getCardProps };
