+++ src/components/search/results.tsx
@@ -1,35 +1,34 @@
-import { TypographyScale } from '@mdb/flora';
-
+import React from 'react';
 import { ResultsProps } from './types';
 import { dataStyles } from './styles';
 
-const Results: React.FunctionComponent<ResultsProps> = ({
-    data,
-    isLoading,
-    hasError,
-}) => {
-    return (
-        <div sx={dataStyles}>
-            {data &&
-                data.map(page =>
-                    page.map(piece => (
-                        <div sx={{ width: '100%' }} key={piece.name}>
-                            <TypographyScale variant="heading6">
-                                {piece.name}
-                            </TypographyScale>
-                            <TypographyScale variant="body2">
-                                {piece.description}
-                            </TypographyScale>
-                        </div>
-                    ))
-                )}
-            {isLoading ? (
-                <div>Loading...</div>
-            ) : hasError ? (
-                <div>Something went wrong, please try again</div>
-            ) : null}
-        </div>
-    );
-};
+import Card, { getCardProps } from '../card';
+
+const Results: React.FunctionComponent<ResultsProps> = React.memo(
+    ({ data, isLoading, hasError }) => {
+        console.log('Results Rendering');
+        return (
+            <div sx={dataStyles}>
+                {data &&
+                    data.map(page =>
+                        page.map(piece => (
+                            <Card
+                                key={piece.slug}
+                                sx={{ width: '100%' }}
+                                {...getCardProps(piece, 'list')}
+                            />
+                        ))
+                    )}
+                {isLoading ? (
+                    <div>Loading...</div>
+                ) : hasError ? (
+                    <div>Something went wrong, please try again</div>
+                ) : null}
+            </div>
+        );
+    }
+);
+
+Results.displayName = 'Results';
 
 export default Results;
