+++ src/components/search/search.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import dynamic from 'next/dynamic';
 import useSWRInfinite from 'swr/infinite';
 
 import {
@@ -12,7 +13,6 @@ import {
 
 import { ContentPiece } from '../../interfaces/content-piece';
 
-import Results from './results';
 import {
     titleStyles,
     searchBoxStyles,
@@ -22,6 +22,10 @@ import {
 import { SearchProps, SortByType } from './types';
 import { fetcher, sortByOptions } from './utils';
 
+const DynamicResults = dynamic(() => import('./results'), {
+    loading: () => <p>...</p>,
+});
+
 const Search: React.FunctionComponent<SearchProps> = ({
     name,
     hideSortBy = false,
@@ -29,6 +33,7 @@ const Search: React.FunctionComponent<SearchProps> = ({
 }) => {
     const [search, setSearch] = useState('');
     const [sortBy, setSortBy] = useState<SortByType>('recent');
+    const [showResults, setShowResults] = useState(false);
 
     const getKey = (pageIndex: number, previousPageData: ContentPiece[]) => {
         if (previousPageData && !previousPageData.length) return null;
@@ -50,10 +55,12 @@ const Search: React.FunctionComponent<SearchProps> = ({
     const onSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
         setSize(1);
         setSearch(event.target.value);
+        setShowResults(true);
     };
     const onSort = (val: string) => {
         setSize(1);
         setSortBy(sortByOptions[val] as SortByType);
+        setShowResults(true);
     };
 
     return (
@@ -82,19 +89,25 @@ const Search: React.FunctionComponent<SearchProps> = ({
                         width="100%"
                     />
                 )}
-                <Results
-                    data={data}
-                    isLoading={isValidating}
-                    hasError={error}
-                />
-                <div sx={loadMoreStyles}>
-                    <Button
-                        onClick={() => setSize(size + 1)}
-                        variant="secondary"
-                    >
-                        Load more
-                    </Button>
-                </div>
+                {showResults && (
+                    <>
+                        <DynamicResults
+                            data={data}
+                            isLoading={isValidating}
+                            hasError={error}
+                        />
+                        <div sx={loadMoreStyles}>
+                            {!isValidating && data && (
+                                <Button
+                                    onClick={() => setSize(size + 1)}
+                                    variant="secondary"
+                                >
+                                    Load more
+                                </Button>
+                            )}
+                        </div>
+                    </>
+                )}
             </GridLayout>
         </form>
     );
