--- src/constants/content-categories.ts
@@ -1,8 +1,0 @@
-import { PillCategory } from '../types/pill-category';
-
-export const CONTENT_CATEGORIES: PillCategory[] = [
-    'ARTICLE',
-    'VIDEO',
-    'DEMO APP',
-    'TUTORIAL',
-];
